Initial commit for removing SAML 1.x implementations.
