+++ modules/rampart-trust/src/main/java/org/apache/rahas/RahasData.java
@@ -29,7 +29,7 @@ import org.apache.ws.security.handler.WSHandlerConstants;
 import org.apache.ws.security.handler.WSHandlerResult;
 import org.apache.ws.security.KerberosTokenPrincipal;
 import org.apache.ws.security.message.token.SecurityTokenReference;
-import org.opensaml.SAMLAssertion;
+//import org.opensaml.SAMLAssertion;
 import org.opensaml.saml2.core.Assertion;
 import org.w3c.dom.Element;
 
@@ -87,7 +87,7 @@ public class RahasData {
     
     private String  claimDialect;
     
-    private SAMLAssertion assertion;
+//    private SAMLAssertion assertion;
     
     private Assertion saml2Assertion;
 
@@ -203,8 +203,8 @@ public class RahasData {
 							this.saml2Assertion = (Assertion) wser
 									.get(WSSecurityEngineResult.TAG_SAML_ASSERTION);
 						} else {
-							this.assertion = (SAMLAssertion) wser
-									.get(WSSecurityEngineResult.TAG_SAML_ASSERTION);
+//							this.assertion = (SAMLAssertion) wser
+//									.get(WSSecurityEngineResult.TAG_SAML_ASSERTION);
 						}
 					} else if (act == WSConstants.KERBEROS || act == WSConstants.KERBEROS_SIGN) {
 						this.principal = (KerberosTokenPrincipal) principalObject;
@@ -214,7 +214,7 @@ public class RahasData {
                 }
             }
 			// If the principal or a SAML assertion is missing
-			if (this.principal == null && (this.assertion == null || saml2Assertion == null)) {
+			if (this.principal == null && (saml2Assertion == null)) {
 				throw new TrustException(TrustException.REQUEST_FAILED);
 			}
         }
