+++ modules/rampart-trust/src/main/java/org/apache/rahas/impl/SAML2TokenRenewer.java
@@ -5,6 +5,7 @@ import org.apache.axiom.om.OMNode;
 import org.apache.axiom.om.impl.dom.jaxp.DocumentBuilderFactoryImpl;
 import org.apache.axiom.soap.SOAPEnvelope;
 import org.apache.axis2.context.MessageContext;
+import org.apache.axis2.description.Parameter;
 import org.apache.commons.lang.StringUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
@@ -13,6 +14,7 @@ import org.apache.rahas.impl.util.SAMLUtils;
 import org.apache.rahas.impl.util.SignKeyHolder;
 import org.apache.ws.security.WSSecurityException;
 import org.apache.ws.security.components.crypto.Crypto;
+import org.apache.ws.security.components.crypto.CryptoFactory;
 import org.apache.ws.security.saml.SAMLUtil;
 import org.apache.ws.security.util.XmlSchemaDateFormat;
 import org.apache.xml.security.c14n.Canonicalizer;
@@ -50,7 +52,13 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
-public class SAML2TokenRenewer extends SAMLTokenRenewer implements TokenRenewer {
+public class SAML2TokenRenewer implements TokenRenewer {
+
+    private String configParamName;
+
+    private OMElement configElement;
+
+    private String configFile;
 
     protected List<Signature> signatureList = new ArrayList<Signature>();
 
@@ -231,5 +239,117 @@ public class SAML2TokenRenewer extends SAMLTokenRenewer implements TokenRenewer 
         }
         return builder.buildObject(objectQName.getNamespaceURI(), objectQName.getLocalPart(),
                 objectQName.getPrefix());
+    }
+
+    /**
+     * set the configuration for SAML 1.1 and 2.0 renewing
+     * @param inMsgCtx
+     * @return
+     * @throws TrustException
+     */
+    protected SAMLTokenIssuerConfig  setConfig(MessageContext inMsgCtx) throws TrustException {
+        SAMLTokenIssuerConfig config = null;
+        if (this.configElement != null) {
+            config = new SAMLTokenIssuerConfig(configElement
+                    .getFirstChildWithName(SAMLTokenIssuerConfig.SAML_ISSUER_CONFIG));
+        }
+        // Look for the file
+        if (config == null && this.configFile != null) {
+            config = new SAMLTokenIssuerConfig(this.configFile);
+        }
+        // Look for the param
+        if (config == null && this.configParamName != null) {
+            Parameter param = inMsgCtx.getParameter(this.configParamName);
+            if (param != null && param.getParameterElement() != null) {
+                config = new SAMLTokenIssuerConfig(param
+                        .getParameterElement().getFirstChildWithName(
+                                SAMLTokenIssuerConfig.SAML_ISSUER_CONFIG));
+            } else {
+                throw new TrustException("expectedParameterMissing",
+                        new String[] { this.configParamName });
+            }
+        }
+        if (config == null) {
+            throw new TrustException("configurationIsNull");
+        }
+        if(config.isTokenStoreDisabled()){
+            throw new TrustException("errorTokenStoreDisabled");
+        }
+        //initialize and set token persister and config in configuration context.
+        if (TokenIssuerUtil.isPersisterConfigured(config)) {
+            TokenIssuerUtil.manageTokenPersistenceSettings(config, inMsgCtx);
+        }
+        return config;
+    }
+
+    /**
+     * create the RSTR element with the token type
+     * @param inMsgCtx
+     * @param data
+     * @param env
+     * @param tokenType
+     * @return
+     * @throws TrustException
+     */
+    protected OMElement buildResponse(MessageContext inMsgCtx, RahasData data, SOAPEnvelope env, String tokenType) throws TrustException {
+        // Create RSTR element, with respective version
+        OMElement rstrElem;
+        int wstVersion = data.getVersion();
+        if (RahasConstants.VERSION_05_02 == wstVersion) {
+            rstrElem = TrustUtil.createRequestSecurityTokenResponseElement(
+                    wstVersion, env.getBody());
+        } else {
+            OMElement rstrcElem = TrustUtil
+                    .createRequestSecurityTokenResponseCollectionElement(
+                            wstVersion, env.getBody());
+            rstrElem = TrustUtil.createRequestSecurityTokenResponseElement(
+                    wstVersion, rstrcElem);
+        }
+        // Create TokenType element
+        TrustUtil.createTokenTypeElement(wstVersion, rstrElem).setText(
+                tokenType);
+        return rstrElem;
+    }
+
+    /**
+     * create the crypto from configuration. Used in SAML2tokenRenewer as well
+     * @param inMsgCtx
+     * @param config
+     * @return
+     */
+    protected Crypto getCrypto(MessageContext inMsgCtx, SAMLTokenIssuerConfig config){
+        Crypto crypto;
+        if (config.cryptoElement != null) {
+            // crypto props defined as elements
+            crypto = CryptoFactory.getInstance(TrustUtil
+                    .toProperties(config.cryptoElement), inMsgCtx
+                    .getAxisService().getClassLoader());
+        } else {
+            // crypto props defined in a properties file
+            crypto = CryptoFactory.getInstance(config.cryptoPropertiesFile,
+                    inMsgCtx.getAxisService().getClassLoader());
+        }
+        return crypto;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public void setConfigurationFile(String configFile) {
+        this.configFile = configFile;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public void setConfigurationElement(OMElement configElement) {
+        this.configElement = configElement;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public void setConfigurationParamName(String configParamName) {
+        this.configParamName = configParamName;
     }
 }
