+++ orbit/rampart-core/pom.xml
@@ -139,10 +139,10 @@
             <groupId>org.apache.axis2</groupId>
             <artifactId>axis2-kernel</artifactId>
         </dependency>
-        <dependency>
-            <groupId>org.opensaml</groupId>
-            <artifactId>opensaml1</artifactId>
-        </dependency>
+<!--        <dependency>-->
+<!--            <groupId>org.opensaml</groupId>-->
+<!--            <artifactId>opensaml1</artifactId>-->
+<!--        </dependency>-->
         <dependency>
             <groupId>org.opensaml</groupId>
             <artifactId>opensaml</artifactId>
