+++ pom.xml
@@ -418,11 +418,11 @@
                 <artifactId>commons-lang</artifactId>
                 <version>${commons-lang.version}</version>
             </dependency>
-            <dependency>
-                <groupId>org.opensaml</groupId>
-                <artifactId>opensaml1</artifactId>
-                <version>${opensaml1.version}</version>
-            </dependency>
+<!--            <dependency>-->
+<!--                <groupId>org.opensaml</groupId>-->
+<!--                <artifactId>opensaml1</artifactId>-->
+<!--                <version>${opensaml1.version}</version>-->
+<!--            </dependency>-->
             <dependency>
                 <groupId>org.apache.ws.security.wso2</groupId>
                 <artifactId>wss4j</artifactId>
@@ -552,7 +552,7 @@
         <neethi.version>2.0.4.wso2v5</neethi.version>
         <opensaml.version>2.6.6</opensaml.version>
         <opensaml.xmltooling.version>1.4.6</opensaml.xmltooling.version>
-        <opensaml1.version>1.1</opensaml1.version>
+<!--        <opensaml1.version>1.1</opensaml1.version>-->
         <commons.ssl.version>0.3.9</commons.ssl.version>
 
         <junit.version>4.12</junit.version>
